+++ ext/MarkdownExtension.jl
@@ -7,7 +7,7 @@ using Typstry: TypstContext, compile_workload, show_raw
 
 show_typst(io::IO, tc::TypstContext, x::MD) = show_raw(io, tc, MIME"text/markdown"(), :markdown, x)
 
-const examples = [md"# A" => MD => [:backticks, :block, :depth, :indent, :mode]]
+const examples = [md"# A" => MD => [:block, :depth, :lang, :align, :syntaxes, :theme, :tab_size]]
 
 compile_workload(examples)