+++ src/strings/Typsts.jl
@@ -20,7 +20,7 @@ whose [`show_typst`](@ref) method formats the wrapped value.
 - `repr(::MIME"text/typst\u00A0::Typst; context = nothing)`
 - `show_typst(::IO,\u00A0::TypstContext,\u00A0::Typst)`
 - `show(::IO,\u00A0::MIME"text/typst",\u00A0::Typst)`
-    - Accepts `IOContext(::IO,\u00A0:typst_context\u00A0=>\u00A0::TypstContext)`
+    - Accepts `IOContext(::IO,\u00A0::TypstContext)`
 - `show(::IO,\u00A0::Union{MIME"application/pdf",\u00A0MIME"image/png",\u00A0MIME"image/svg+xml"},\u00A0::Typst)`
     - Accepts `IOContext(::IO,\u00A0::TypstContext)`
     - Uses the `preamble` in [`context`](@ref Typstry.context)
@@ -29,10 +29,7 @@ whose [`show_typst`](@ref) method formats the wrapped value.
 # Examples
 
 ```jldoctest
-julia> t = Typst(1)
-Typst{Int64}(1)
-
-julia> show(stdout, "text/typst", t)
+julia> show_typst(Typst(1))
 \$1\$
 ```
 """