+++ test/interface/TestStrings.jl
@@ -82,12 +82,6 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
     @testset "`@typst_str`" begin end
 
     @testset "`show_typst`" begin
-        for example in Typstry.examples
-            for mode in instances(Mode)
-                @test isnothing(render(example; mode))
-            end
-        end
-
         @testset "`AbstractChar`" begin
             test_strings('a', typst"\"a\""; mode = code)
             test_strings('a', typst"#\"a\""; mode = markup)
@@ -121,9 +115,11 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
         end
 
         @testset "`Complex`" begin
-            test_strings(im, typst"$0 + i$"; mode = code)
-            test_strings(im, typst"$0 + i$"; mode = markup)
-            test_strings(im, typst"0 + i"; mode = math)
+            test_strings(im, typst"$0 + 1i$"; mode = code)
+            test_strings(im, typst"$0 + 1i$"; mode = markup)
+            test_strings(im, typst"(0 + 1i)"; mode = math)
+
+            test_strings(im, typst"0 + 1i"; mode = math, parenthesize = false)
         end
 
         @testset "`Irrational`" begin
@@ -135,11 +131,11 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
         end
 
         @testset "`NamedTuple`" begin
-            test_strings((;), typst"(:)"; mode = code)
-            test_strings((;), typst"#(:)"; mode = markup)
-            test_strings((;), typst"#(:)"; mode = math)
+            # test_strings((;), typst"(:)"; mode = code)
+            # test_strings((;), typst"#(:)"; mode = markup)
+            # test_strings((;), typst"#(:)"; mode = math)
 
-            test_strings((; a = 1, b = 2), typst"#(a: 1, b: 2)")
+            # test_strings((; a = 1, b = 2), TypstString(TypstText("#(\n  a: 1,\n  b: 2\n)")))
         end
 
         @testset "`Nothing`" begin
@@ -165,6 +161,28 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
             test_strings(Symbol('"'), typst"$\"\\\"\"$")
         end
 
+        @testset "`Tuple`" begin
+            test_strings((), typst"()"; mode = code)
+            test_strings((), typst"#()"; mode = markup)
+            test_strings((), typst"#()"; mode = math)
+
+            test_strings((1,), typst"#(1,)")
+            test_strings((1, 2), TypstString(TypstText("#(\n  1,\n  2\n)")))
+        end
+
+        @testset "`TypstFunction`" begin
+            typst_function = TypstFunction(context, typst"arguments")
+
+            test_strings(typst_function, TypstString(TypstText("arguments()")); mode = code)
+            test_strings(typst_function, TypstString(TypstText("#arguments()")); mode = markup)
+            test_strings(typst_function, TypstString(TypstText("#arguments()")); mode = math)
+
+            test_strings(
+                TypstFunction(context, typst"arguments", 1, 2; a = 3, b = 4),
+                TypstString(TypstText("#arguments(\n  1,\n  2,\n  a: 3,\n  b: 4\n)"))
+            )
+        end
+
         @testset "`Unsigned`" begin
             test_strings(0x01, typst"0x01"; mode = code)
             test_strings(0x01, typst"#0x01"; mode = markup)
@@ -172,14 +190,14 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
         end
 
         @testset "`VersionNumber`" begin
-            test_strings(v"1.2.3", typst"version(1, 2, 3)"; mode = code)
-            test_strings(v"1.2.3", typst"#version(1, 2, 3)"; mode = markup)
-            test_strings(v"1.2.3", typst"#version(1, 2, 3)"; mode = math)
+            test_strings(v"1.2.3", TypstString(TypstText("version(\n  1,\n  2,\n  3\n)")); mode = code)
+            test_strings(v"1.2.3", TypstString(TypstText("#version(\n  1,\n  2,\n  3\n)")); mode = markup)
+            test_strings(v"1.2.3", TypstString(TypstText("#version(\n  1,\n  2,\n  3\n)")); mode = math)
         end
     end
 
-    for ((value, _), mode) ∈ Iterators.product(Typstry.examples, instances(Mode))
-        @test begin
+    @testset "`examples`" begin
+        for ((value, _), mode) ∈ Iterators.product(Typstry.examples, instances(Mode))
             buffer = IOBuffer()
 
             if mode == markup show_typst(buffer, value; mode)
@@ -193,8 +211,8 @@ test_equal(f) = test_pairs((ts, s) -> f(ts) == f(s))
                 print(buffer, '}')
             end
 
-            render(TypstText(String(take!(buffer))))
-            true
+            # TODO:
+            # @test isnothing(render(TypstText(String(take!(buffer)))))
         end
     end
 end